requirements.txt
Flask-MySQLdb==0.2.0
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.20


